fix: We can't use short format of mango query
Sift doesn't support the sort format that mango does.
```
where({ worker: ['client', 'konnector']})
```

is translated to an `$and` by mango, but sift doesn't
do that, so we have issues when some query where emptyed
because of that.

The fix here is to throw an error if we see this usage.

BREAKING CHANGE: if you use a selector on multiple value
without the `$and`  mango operator, you have to change that
and use this `$and` operator.