feat: We don't want to concatenate `additionalName` in the `displayName`
This attribute will be used within cozy-contacts.

In this case, we do not want to concatenate in the
`displayName` attribute the `additionalName` attribute.

A simple solution is to offer the possibility to `makeFullname`
to construct the `displayName` in a personalized way
(within the limits of the attributes existing in `name`)