feat: Add `getSetting` and `saveAfterFetchSetting` methods to CozyClient
In our cozy-apps, we often need to handle specific settings for each
cozy-app. This includes:
- io.cozy.settings/io.cozy.settings.instance
- io.cozy.settings/io.cozy.settings.bitwarden
- io.cozy.home.settings
- io.cozy.coachco2.settings
- io.cozy.mespapiers.settings

In the current code base, all those access are made with different code
that can be heterogeneous

We want to mutualize those access into two methods:
- client.getSetting(slug, key)
- client.saveAfterFetchSetting(slug, key, value)

`getSetting` is responsible to retrieve the given `key` from a specific
cozy-app setting

`saveAfterFetchSetting` is responsible to update the given `key` for a
specific cozy-app setting

Note that `saveAfterFetchSetting` will first fetch the cozy-app
settings before editing and saving them. This prevents from having to
pass existing settings as parameter and also this ensure we always
inject up to date data into the database