refactor(pouch-link): Make localStorage a class an return async methods
We want to reduce coupling between CozyPouchLink and the browser's
local storage

The first step is to groupe existing methods into a PouchLocalStorage
class, this will allow (in the next commit) to inject a storage adapter
into the class constructor

Also in some platforms like ReactNative, the localstorage methods are
asynchronous. So in order to be compatible with all platforms we want
those methods to become async by default

The big main of making those methods async is that we cannot call them
from the `PouchManager` constructor. So we introduce a new `.init()`
method in the `PouchManager` class. This async method now contains all
the initialization logic and should be called after creating a new
`PouchManager`

As `PouchManager` is called internally by `CozyPouchLink` and is not
meant to be a publicly available class, then we don't consider this as
a breaking change