feat: Persist documents to local Pouch when they have no `rev`
When a PouchLink is configured, the synchronization process would
retrieve only documents that exists on the remote database

Those documents would be served as usual when a `.query()` is made
using the PouchLink

Unfortunately, not all `.query()` correspond to a document that exists
on the remote database. With the collection mechanism, some documents
are retrieved using an alternative endpoint and then the cozy-stack
generate the document on-demand

For example this is the case for `io.cozy.settings.flags`,
`io.cozy.settings.context` and `io.cozy.settings.disk-usage`

Because those documents does not actually exist on the remote database,
we cannot sync them using the PouchDB sync mechanism and so they won't
be available when the device is offline

To fix this, we chose to manually insert those document in the local
PouchDB when we detect that a `.query()` call retrieve them from the
cozy-stack

The easiest way we found to detect them is to check for the existence
of `meta.rev` object in the `.query()` result. If no `rev` is present,
this means that the documents does not come from the remote database
but has been dynamically generated by the cozy-stack

Those documents are inserted into existing PouchBD with the
`cozyLocalOnly` attributes, so we know that they should not be
replicated to the remote database on next sync

In order to make the persist mechanism possible, we implemented a new
`persistData()` method to the CozyLink interface

The impact is that every `Link` should now implement the
`persistData()` interface that should persist the data if it is able to
do it, or to forward the action otherwise

If no `Link` process the action, then the final "default" method will
do nothing (contrary to the default for `request()` that would throw an
error)

BREAKING CHANGE: CozyClient's links should now implement a
`persistData()` method that should persist the given data if it is able
to do it, or to forward the action to the next `Link` otherwise