fix(nextcloud): Allow folder query to be re-evaluated
When cozy-client receives a request, it re-evaluates the other
requests in memory to ensure that they are up to date with the added or
updated documents.

However, this re-evaluation did not work for requests retrieving the
contents of a Nextcloud folder. Unlike `io.cozy.files` which has a
`dir_id`, we use the folder path to retrieve its contents. File paths
include their names, which stops a direct match with the query selector
based on the path to the parent folder. For example, the file path
`/folder1/file1.txt` would not directly match a query looking for all
files in `/folder1`.

To resolve this problem, this commit adds a `parentPath` property to
files, corresponding to the path of their parent folder. This attribute
allows the path comparison to correctly identify files located in the
same parent folder, even if their names vary.

So, after moving files, we can query the contents of the destination
folder. Files that have been moved are automatically removed from the
query in the source folder, ensuring consistency in the interface and
the data displayed.