feat:  Reset a query based on its id
The resetQuery method can be used to reset a query in the cache to its
initial state based on its id. The query will refetch after.

This can be useful after a mutation without response. Because you know
that some of the data in a query is no longer fresh, you can choose to
update them before fetchPolicy expires.