fix: Do not persist virtual documents when served from CozyPouchLink
When querying documents from CozyPouchLink, then the `meta.rev`
attribute is not added to the document. This happens because this
attributes does not exist in database but is added by the cozy-stack
when serving JSON:API format

We don't want to add the `meta.rev` attribute by ourself as it may have
side effects on the synchronization process (i.e. adding `meta.rev` in
the database documents)

Instead we want to introduce a `cozyFromPouch` attribute that is added
by CozyPouchLink when returning the result