feat: Do not fail replication when some databases do no exist
It would be tempting to fail the replication when one of the databases
does not exist as it is not an expected behavior

But is some situation it is expected. For example this is the case for
io.cozy.accounts that does not exist on the server side until a
Konnector is installed and configured

So we prefer to have failing attempts on each try instead of blocking
the entire replication for that kind of error

BREAKING CHANGE: cozy-pouch-link now uses AggregateError class. So the
library should be use on platforms that handle this class
(i.e. ReactNative 72+, Node 15+)

More info:
https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/AggregateError