feat: Remove the implicit inside partialFilter to find existing index
When CouchDB creates an index with a partialFilter, it adds explicit
operators when they are implicit, typically the $and and $eq operators.
This causes a mismatch when comparing the partialFilter from the
request definition with the partialFilter from the index. To address
this, we added a step to make the operators in the partialFilter from
the request definition explicit.

This makes it possible to migrate indexes after they have been
renamed rather than re-creating them, which has a non-negligible cost.