feat: Discriminate not_synced doctypes from ones with non-blocking error
With current implementation, the CozyPouchLink would not process a
query for a doctype that did not succeed to synchronize and it would
forward the query to the next link

This is problematic for 2 reasons

First because we now expect the CozyPouchLink to be the last link of
the chain, so forwarding the query would result to an exception thrown

Second because we previously added a mechanism that allows some pouch
to safely fail. This happens when the remote Couch does not contain the
doctype yet (i.e. `io.cozy.accounts` exist only after configuring the
first konnector). In that scenario we don't want the query to throw but
to return an empty list

To improve this, we want to forward the query only when the
synchronization has never been started or when it unexpectedly failed

We now have 3 synchronization status:
- `not_synced` when synchronization never started or after an
  unexpected error
- `not_complete` when synchronization occurred but failed with a
  non-blocking error
- `synced` when synchronization succeeded