fix(move): Trash the conflicting file instead of the destination folder
The folder was passed directly to the destination, so the path was
always present. This had the effect of recovering the folder and moving
it to the trash. With this fix, the destination path is calculated
correctly in order to find the file at the origin of the conflict and
move it to the trash.