feat: Increase PouchdDB default batch size for view updates
The `view_update_changes_batch_size` parameter is used to determine the
docs batch size when computing view updates. PouchDB default is 50,
which is quite low, and badly hurt performancs when dealing with
thousands of docs and the first view computation.

With a 44K files database, in a react-native env:

**Before:**
Indexing time: 53 868 ms

**After:**
Indexing time: 31 833 ms