feat: Do not query /jobs/triggers with a partialFilter
One might want to avoid querying the `/jobs/triggers` route and stick to
the `/data` route, to avoid extra work from the stack, that can be
costful.
We were using the existence of `worker` and `type` keys in selector, but
a partialFilter existence should also force the use of the `/data`
route.