feat: Add PerformanceApi to CozyClient
We want to be able to measure CozyClient's performances in web apps but
also in mobile apps

To make this possible we introduce the PerformanceAPI

This API should be injected in CozyClient's constructor, then
CozyClient will be able to use it to measure performances

This API is meant to wrap platforms' native APIs. On web browsers we
wrap the well known Performance API, but on mobile app we may want to
use the implementation from react-native-performances

The CozyClient provides the web implementation, but other
implementation should be implemented by consuming apps and then
injected in the CozyClient's constructor

If no implementation is injected in the constructor, then the
`defaultPerformance` implementation is used. This implementation does
nothing, so there is no impact on web apps by default

Related link:
https://developer.mozilla.org/en-US/docs/Web/API/Performance