feat: Do not hydrate doc if the relationship does not exist
We used to hydrate any document with the relationships existing in the
provided schema, even though the relationship does not exist on the
document. We now hydrate only if the relationship is set in the
document.
Note we can still force the hydratation through the `autoHydrate`
option.

BREAKING CHANGE: the relationship hydration is made only if the
relationship exists in the document, so the developer should not assume
a `document.relationshipName` is always defined, anymore.
As an alternative, it is possible to pass `autoHydrate` on
cozy-client options to ease migration, e.g:
```
const newClient = new CozyClient({
  schema,
  autoHydrate: true
})
```
However, please not this has performance impact, as it forces
extra-check on store queries evaluation.