feat: Improve React performances by avoiding useless renders
We noticed that any query or mutation resulted in many unexpected render
by the apps, even when there is no new data.  This is particularly
problemetic in environement with low ressources such a the mobile: we
measured strong bottlenecks because of this, making the main thread
overloaded and resulting in not smooth experience for the user.

Each time a i`useQuery` is made, it is saved as a listener in redux,
through the useSelector hook. We provide the result of
`getQueryFromState` to this hook to evaluate whether or not the state
changed. By default, this evaluation is a referential equality check.
See https://react-redux.js.org/api/hooks#useselector

> When an action is dispatched, useSelector() will do a reference
comparison of the previous selector result value and the current result
value. If they are different, the component will be forced to re-render.

The issue here is that the `getQueryFromState` function always return a
new object. Therefore, the reference comparison will never be true.

And because the dispatch will make the evaluation of all listeners, it
means any query will force the re-render to all components having a
useQuery, leading to performance issues.

We now have our own comparison function, taking advantage of our store
specifities. This is passed to the `useSelector` hook to determine
whether or not two queries have changed and if a render should be made.

With this improvement, we made some measures:

- Display home page of Drive
  - [BEFORE] 6899 ms
  - [AFTER] 3590 ms
  - => x1.9 boost
- Upload 100 files of 1K
  - [BEFORE] 107446 ms
  - [AFTER] 18395 ms
  - => x5.8 boost