feat: Allow to enforce Stack link on request chain
In previous commits we implemented supports for offline mode through
CozyPouchLink and FlagshipLink

Since this feature, the StackLink is able to check for internet
reachability by using the `isOnline()` method. When no internet is
detected, then the request is forwarded to the next link (i.e. the
CozyPouchLink)

In some scenario we may want to prevent that behavior. This is the case
when we are working on a feature that cannot work without internet
access, or when the query's result is expected to contain data that is
injected by the cozy-stack (and so that is not available in the local
Pouch database)

To make this possible we introduce the `forceStack` parameter that can
be set in the query options

When set to `true` then the StackLink will not attempt to forward the
request when offline, instead it will still attempt to do the request
and throw

We chose to attempt the request instead of directly throwing an error
in order to prevent hypothetical scenario where `isOnline()` method
returns false even if internet is reachable

BREAKING CHANGE: CozyLink's request methods now takes an additional
`options` argument that can be used to enforce usage of Stack link.
Although `.request()` is meant to be an internal API, if your code
calls it, you'll want to introduce a new argument for `options`. This
is also the case if you instanciate a new CozyLink with a handler
argument for `request`

Before:
```
// Initialization
new CozyLink((operation, result = '', forward) => {
  return forward(operation, result + 'foo')
})

// Call
link.request(operation)

// Call with result and forward
link.request(operation, null, () => { /* do stuff */ })
```

After:
```
// Initialization
new CozyLink((operation, options, result = '', forward) => {
  return forward(operation, options, result + 'foo')
})

// Call
link.request(operation, options)

// Call with result and forward
link.request(operation, options, null, () => { /* do stuff */ })
```