feat: Ensure previous replication are done
If an app wants to trigger replication at different event handling, it
is nice to have some control on it, and prevent multiple parallel
replication processes.
Thus, we introduce a `waitForReplications` param, which is set to true
by default.
Thanks to this, we wait for the previous replication process to complete
before running the new one