fix: Correctly handle document's store updates
We used to create a new reference on the `documents` state for any
query.
This is problematic because we do a referential comparison on the
updated `documents` state to evaluate if there is any update on it,
resulting on all store's queries re-evaluation.
So, basically, we were doing a full queries re-evaluation for *any*
query execution, even if the same query is run several time with
no change.
This is particularly problematic when there are a lot of queries and/or
queries with many documents returned, as their re-evaluation will be
costful. Moreover, the document merging itself is costful.

The comparison between old and new docs is crucial for performances, so
we did some performances measures: we took a worst case where we fill
the store with 25K docs with a query and run again the same query with
no document change.

What we measure here is the time taken in the `extractAndMergeDocument`
method:

- Before: 73892 ms
- Compare with lodash isEqual: 47530 ms
- Compare with stringify + djb2 hash: 19785 ms
- Compare with fast-deep-equal: 2404 ms
- Compare with JSON.stringify: 1899 ms
- Compare with rev (for the record): 123ms

Note the JSON.stringify comparison is faster, but it does not guarantee
key order: 2 documents with same value but different order will be seen
as different. For safety, we prefer fast-deep-equal, slighly slower, but
more reliable.