fix: Correctly handle document's store updates
We used to create a new reference on the `documents` state for any
query.
This is problematic because we do a referential comparison on the
updated `documents` state to evaluate if there is any update on it,
resulting on all store's queries re-evaluation.
So, basically, we were doing a full queries re-evaluation for *any*
query execution, even if the same query is run several time with
no change.
This is particularly problematic when there are a lot of queries and/or
queries with many documents returned, as their re-evaluation will be
costful.