feat: Add queryEngine platform option
This is now possible to pass a custom queryEngine, which is in charge of
querying the database.  The reason of this is to offer the possibility
to a client to query the local database without using Pouchdb API: as we
discovered, using this API can introduce a significant performance
overhead, because of additional treatments that are not always
necessary, but particularly because of the way mango indexes are
handled: each time an index is created, a complete database is created
(called mrview) and the indexed ids and revisions are duplicated in it,
which is very costful (some measures are provided in the next commit).

Instead, a direct database access can be much more efficient, e.g. by
directly accesing SQLite or IndexedDB and using native indexes.