fix: Do not store existing documents
The `persistCozyData` function now avoid to save existing document, i.e.
that already exist on the database with the same fields. We could not
rely on revision here, because some doc might have no revision, or a
changed fields with the same doc revision.
We took this opportunity to make some refactoring and add some
additional checks.