feat: Change PouchDB databases naming
We introduce a `__doctype` separator in the database name, to easily
differenciate between the url part and the doctype part. Before, it was
not posible to extract the doctype from the name, because a single `_`
separator was used, while doctypes can include `_` in their name.  As
PouchDB is not heavily used for now, let's take this opportunity to
introduce this change that will be much more difficult to do in the
future.

WARNING: please note the previous PouchDB databases will be
automatically removed.