feat: Add disableStoreForQueries options to bypass store
When doing a query, one might want to avoid using the redux store.
This can typically be useful for situations where data is already
reachable locally efficiently, or to preserve CPU/RAM when dealing
with a lot of data.
Thus we introduce a `disableStoreForQueries`, which void to rely on
the redux store to evaluate queries or store data results