feat: Use `useCustomStore` rather than `store`
BREAKING CHANGE: to pass a custom store from the app, we used to rely on
the boolean `store` option. The naming is confusing, because we could
interpret it as completely disabling the redux store, which is not the
case, as the aim is to use either the cozy-client internal store, or the
application own store. It was also confusing in the code, as the store
variable is the store itself.
Thus, we rename the option in `useCustomStore` to better reflect the
intention.

If your application was using `store: false` in the cozy-client
instanciation, you should now use `useCustomStore: true`.