feat: Add DataProxyLink
This new link will be useful to allow apps to communicate directly with
the DataProxy by using it as primary link.
Eventually, the objective will be to make all the cozy-client request to
go through this link, so the dataproxy will handle it either locally or
remotely.
Note that DataProxyProvider directly handle web or mobile environments:
it creates an iframe for web apps, while it relies on webview intent
for mobile.