fix: Do not throw when doc has relationship but not schema
When usign the DataProxyLink, queries are made by the app, and executed
in the DataProxy worker.
Some docs might have relationships and queries could have `.includes`
directive.
However, there is no schema on the Dataproxy side, which was causing
errors because relationships declaration are expected in the schema in
such situations.
Thus, we now do an early return if there is no relationsips in the
schema to avoid such throw and expect the cozy-client's app to handle
relationship rather than DataProxy