fix: Do not throw if there is no database for a doctype
At the link instanciation, we declare which doctypes we want to persist.
For some "virtual" doctypes, there is no stack database, so nothing to
replicate. To deal with such cases, the `persistCozyData` is useful to
save document locally after being queried remotely.
However, we might not want to persist a virtual doctype and continue to
query the data remotely.
In such case, a throw occured because we were trying to query a document
on a non-existent database.

So now, we add a check to avoid such throw.