feat: Handle paths locally
The files path are not stored in CouchDB, only the directories have the
path stored.
When querying the stack, it computes on the fly the path for files.
Thus, we need to do the same on the PouchLink to make it transparent for
the web apps.
We implement a cache mechanism to ensure correct performances