feat: Support multiple create and delete
We used to support mutliple updates but not create nor delete.  Thus, we
now introduce the `createAll` and `destroyAll` API methods, to
respectively bulk create and delete a batch of documents.  `updateAll`is
also introduced, which is a simple wrapper for `saveAll`, just like
`createAll`.

:warning: The `createAll` is currently not supported for stackClient
links, but works for pouchLink.

Adding the associated dsl mutations is also useful for apps that want to
handle themselves store actions. For instance, Drive would be able to
dispatch a single mutation action for multiple updates, typically after
the deletion of a directory. This would improve performances, by
avoiding to make a costful dispatch per document.