fix: Filter missing documents from store
We used to return `null` values when trying to access documents from
store.
This can be problematic as we can end up with query result such as
`data: [null]` because the query involves a missing document.

This can typically happen when a document is deleted while it is
involved in a query. If this query is run from a `getQueryFromState`
before the actual query re-evaluation triggered from the realtime
deletion event, then we'll try to access a missing document.