fix: Reference creation through hydratation
To create relationships on a document, one can use the dedicated methods
after document hydratation.
This is notably used when doing `client.save` if the `references`
argument is provided.

However, we changed the hydratation behaviour on the 60.0.0 release: the
documents are no longer hydrated if no relationship is found.
This breaks creation, so we add an option to force the hydratation for
such case.