feat: Queue DataProxy requests when it is not ready
The DataProxy needs some time to initialized and answer requests.
However, apps may ask for query execution before it is ready ; to
prevent this, we had an init mechanism made in the `DataProxyProvider`
to ensure the children are rendered only when the DataProxy is ready.

It works, but if something went wrong, nothing is displayed, which is
confusing and hard to debug for the developer.

So, we change this by always rendering the children on the DataProxy
side, and implement a queue on the `DataProxyLink` to bufferise
requests, waiting for the dataproxy to be ready.

Once it is ready, all the requests in the queue are executed, in FIFO
fashion.