feat: Make `startReplicationDebounce` run immediatly at first call
When `startReplicationDebounce` is called once, the replication will be
executed immediatly. And then, all subsequent calls made in the debounce
interval will be delayed.

It is useful for situations where a single action triggering a
replication, e.g. an realtime update, should not be delayed to
immediatly benefit from local data.