fix: Prevent infinite loop on replication refresh token
When a replication request ends up with a refresh token error, we
manually make the `refreshToken`, and restart the replication.
If the `refreshToken` request is successful, but the new token is not
set for any reason, this will cause an infinite loop.
To prevent this, we now add a simple check on the renewed token.

This behaviour has been seen on a user with a old version of the
dataproxy, that did not include this fix:
https://github.com/cozy/cozy-client/commit/e392131fe84f5c866317e2ad240ac30196755405