feat: Add entrypoints helpers in applications model
We use entrypoints in cozy-home and cozy-bar and we may use them in
more apps. Here I centralize entrypoints helpers in applications
model (because entrypoints are included in the app manifest).

Besides what was already existing in cozy-home, I add also the
management of entrypoints.condition, to enable or disable entrypoints
with flags.